requirements.txt
openai
streamlit
python-dotenv
PyMuPDF
pdfplumber
docx2txt
spacy
stripe
